add error views
