+++ adminpage/sport/templates/errors/500.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 {% load static %}
 
-{% block title %}{{ error_code }}{% endblock %}
+{% block title %}500{% endblock %}
 
 {%block extrahead %}
 <link rel="stylesheet" href="{% static "sport/css/error.css" %}">
@@ -16,15 +16,11 @@
         </div>
         <div class="col-6 col-right">
             <div class="header">
-                <h1 class="display-1 text-center hdr">{{ error_code }}</h1>
+                <h1 class="display-1 text-center hdr">500</h1>
             </div>
             <div class="body">
                 <div class="col">
                     <p class="display-4 text-center msg">I am sorry, these cowboy-coders made some mistake</p>
-                    <details>
-                        <summary>Details</summary>
-                        <div class="text-left traceback">{% for line in data %}<br/>{{ line }}{% endfor %}</div>
-                    </details>
                 </div>
             </div>
         </div>
@@ -34,12 +30,8 @@
 <div class="container-fluid portrait pl-0 pr-0">
     <div class="col">
         <div class="row justify-content-center">
-            <h1 class="display-1 text-center">{{ error_code }}</h1>
+            <h1 class="display-1 text-center">500</h1>
             <p class="display-4 text-center">I am sorry, these cowboy-coders made some mistake</p>
-            <details>
-                <summary>Details</summary>
-                <div class="text-left traceback">{% for line in data %}<br/>{{ line }}{% endfor %}</div>
-            </details>
         </div>
         <div class="row">
             <img class="align-center" src="{% static "sport/images/errors/error.png" %}">
