+++ adminpage/sport/views/__init__.py
@@ -1,0 +1,3 @@
+from .profile import *
+from .category import *
+from .calendar import *
