+++ adminpage/sport/views/profile.py
@@ -1,12 +1,8 @@
-from typing import Optional
-
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import render, get_object_or_404
+from django.shortcuts import render
 from django.utils import timezone
 
-from api.crud import get_ongoing_semester, get_student_groups, get_trainer_groups, get_brief_hours, get_sports, \
-    get_clubs, get_sc_training_group
-from sport.models import Student, Sport, Trainer
+from api.crud import get_ongoing_semester, get_student_groups, get_brief_hours, get_trainer_groups
 
 
 def parse_group(group: dict) -> dict:
@@ -70,30 +66,3 @@ def profile_view(request, **kwargs):
         })
 
     return render(request, "profile.html", context)
-
-
-@login_required
-def category_view(request, **kwargs):
-    sports = get_sports()
-    clubs = sorted(
-        [{
-            "available_places": club["capacity"] - club["current_load"],
-            **club,
-        }
-            for club in get_clubs()],
-        key=lambda group: (group["current_load"] >= group["capacity"], group["name"])
-    )
-    sc_training_group_id = get_sc_training_group()["id"]
-    return render(request, "category.html", {
-        "sports": sports,
-        "clubs": clubs,
-        "sc_training_group_id": sc_training_group_id,
-    })
-
-
-@login_required
-def calendar_view(request, sport_id, **kwargs):
-    sport = get_object_or_404(Sport, pk=sport_id)
-    return render(request, "calendar.html", {
-        "sport": sport,
-    })
